Fix qwen_image_edit_instyle loader: use bfloat16 as default dtype

+++ qwen_image_edit_instyle/pytorch/loader.py
@@ -75,7 +75,7 @@ class ModelLoader(ForgeModel):
         Returns:
             QwenImageTransformer2DModel with LoRA weights fused.
         """
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         pipe = QwenImageEditPipeline.from_pretrained(
             BASE_MODEL,
@@ -96,7 +96,7 @@ class ModelLoader(ForgeModel):
 
         Returns a dict matching QwenImageTransformer2DModel.forward() signature.
         """
-        dtype = kwargs.get("dtype_override", torch.float32)
+        dtype = kwargs.get("dtype_override", torch.bfloat16)
         batch_size = kwargs.get("batch_size", 1)
 
         # From model config: in_channels=64 (img_in linear input dimension)
